Extract element execution helpers into trait
